requirements.txt
torch
openai-whisper
pydub
ffmpeg-python
tqdm
flask
werkzeug

